continuing

+++ figureout/src/main/java/com/project/figureout/dto/ProductDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter @Setter
@@ -24,13 +25,13 @@ public class ProductDTO {
 
     private float length;
 
-    private double purchaseAmount;
+    private BigDecimal purchaseAmount;
 
     private List<Long> categoriesIds;
 
     private long pricingGroup;
 
-    private double price;
+    private BigDecimal price;
 
     private MultipartFile productImage;
